feat(21-01): move components into Atomic Design directories
- Move 7 atoms: badge, button, input, separator,
  copy-button, toggle, toggle-interactive
- Move 8 molecules: card, callout, code-block, table,
  tabs, tabs-interactive, tooltip, tooltip-interactive
- Move 6 organisms: accordion, accordion-interactive,
  dialog, dialog-interactive, dropdown, dropdown-interactive
- Update barrel exports in src/index.ts
- Fix cross-component and mdx import paths
