+++ packages/artax-ui/src/components/atoms/badge/badge.tsx
@@ -1,7 +1,7 @@
 // ABOUTME: Terminal-styled Badge with monospace font and amber/outline/secondary variants.
 // ABOUTME: Uses cva for variant props and cn() for class merging.
 import { cva, type VariantProps } from 'class-variance-authority'
-import { cn } from '../lib/utils'
+import { cn } from '../../../lib/utils'
 
 const badgeVariants = cva(
   'inline-flex items-center px-2 py-0.5 font-mono text-xs transition-colors',
