+++ packages/artax-ui/src/components/atoms/button/button.tsx
@@ -1,7 +1,7 @@
 // ABOUTME: Terminal-styled Button with $ command prefix and [bracket] ghost variant.
 // ABOUTME: Uses cva for variant/size props and cn() for class merging.
 import { cva, type VariantProps } from 'class-variance-authority'
-import { cn } from '../lib/utils'
+import { cn } from '../../../lib/utils'
 
 const buttonVariants = cva(
   'inline-flex items-center justify-center font-mono text-sm transition-colors focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-amber-accent disabled:pointer-events-none disabled:opacity-50',
