+++ packages/artax-ui/src/components/organisms/accordion/accordion.tsx
@@ -1,6 +1,6 @@
 // ABOUTME: Server-safe Accordion structure components with terminal aesthetic.
 // ABOUTME: Visual shells for accordion layout; interactivity added by accordion-interactive.tsx.
-import { cn } from '../lib/utils'
+import { cn } from '../../../lib/utils'
 
 function AccordionItem({
   className,
